+++ mocks/mock-data.tsv
@@ -1,2 +1,2 @@
-Beautiful & luxurious studio at great location	A quiet cozy and picturesque that hides behind a a river by the unique lightness of Amsterdam.	2023-03-03T08:45:40.284Z	Amsterdam	amsterdam.jpg	amsterdam-01.jpg;amsterdam-02.jpg;amsterdam-03.jpg;amsterdam-04.jpg;amsterdam-05.jpg;amsterdam-06.jpg	true	false	 4.8	apartment	3	4	120	Breakfast	Angelina	angelina@mail.ru	avatar-angelina.jpg	true	52.370216	4.895168
-The house among olive	Peaceful studio in the most wanted area in town	2023-04-06T10:45:40.283Z	Dusseldorf	dusseldorf.jpg	dusseldorf-01.jpg;dusseldorf-02.jpg;dusseldorf-03.jpg;dusseldorf-04.jpg;dusseldorf-05.jpg;dusseldorf-06.jpg	false	true	3.9	room	1	2	225	Washer;Air conditioning	Angelina	angelina@mail.ru	avatar-angelina.jpg	false	51.225402	6.776314
+Beautiful & luxurious studio at great location	A quiet cozy and picturesque that hides behind a a river by the unique lightness of Amsterdam.	2023-03-03T08:45:40.284Z	Amsterdam	amsterdam.jpg	amsterdam-01.jpg;amsterdam-02.jpg;amsterdam-03.jpg;amsterdam-04.jpg;amsterdam-05.jpg;amsterdam-06.jpg	true	false	 4.8	apartment	3	4	120	Breakfast	Angelina	angelina@mail.ru	avatar-angelina.jpg	regular	52.370216	4.895168
+The house among olive	Peaceful studio in the most wanted area in town	2023-04-06T10:45:40.283Z	Dusseldorf	dusseldorf.jpg	dusseldorf-01.jpg;dusseldorf-02.jpg;dusseldorf-03.jpg;dusseldorf-04.jpg;dusseldorf-05.jpg;dusseldorf-06.jpg	false	true	3.9	room	1	2	225	Washer;Air conditioning	Angelina	angelina@mail.ru	avatar-angelina.jpg	pro	51.225402	6.776314
