+++ nodemon.json
@@ -1,0 +1,7 @@
+{
+  "watch": [
+    "src"
+  ],
+  "ext": "ts, json",
+  "exec": "npm run ts ./src/main.ts | pino-pretty --colorize --translateTime SYS:standard"
+}
