+++ src/app/application.ts
@@ -1,9 +1,10 @@
 import { ConfigSchema } from '../core/config/config.schema';
 import { getMongoURI } from '../core/helpers/db.js';
+import { LoggerInfoMessage } from '../core/logger/logger.constants';
+import { AppComponent } from '../types/app-component.enum';
 import { ConfigInterface } from '../types/core/config.interface';
 import { DatabaseClientInterface } from '../types/core/database-client.interface';
 import { LoggerInterface } from '../types/core/logger.interface';
-import { AppComponent, LoggerInfoMessage } from '../utils/constants.js';
 import { inject, injectable } from 'inversify';
 
 @injectable()
