+++ src/core/config/config.service.ts
@@ -1,9 +1,11 @@
 import { ConfigInterface } from '../../types/core/config.interface.js';
 import { LoggerInterface } from '../../types/core/logger.interface.js';
 import { config } from 'dotenv';
-import { AppComponent, ErrorMessage, LoggerInfoMessage } from '../../utils/constants.js';
+import { ErrorMessage } from '../../utils/constants.js';
 import { ConfigSchema, configSchema } from './config.schema.js';
 import { inject, injectable } from 'inversify';
+import { AppComponent } from '../../types/app-component.enum.js';
+import { LoggerInfoMessage } from '../logger/logger.constants.js';
 
 @injectable()
 export default class ConfigService implements ConfigInterface<ConfigSchema> {
