+++ src/core/modules/user/dto/create-user.dto.ts
@@ -2,6 +2,6 @@ export default class CreateUserDto {
   public email!: string;
   public avatarUrl!: string;
   public name!: string;
-  public isPro!: boolean;
+  public userStatus!: string;
   public password!: string;
 }
