+++ src/core/modules/user/user.constants.ts
@@ -1,0 +1,14 @@
+export const UserNameLength = {
+  Min: 1,
+  Max: 15,
+};
+
+export const UserPasswordLength = {
+  Min: 6,
+  Max: 12,
+};
+
+export enum UserStatus{
+  Regular='regular',
+  Pro='pro'
+}
