+++ src/core/modules/user/user.container.ts
@@ -1,9 +1,9 @@
 import { Container } from 'inversify';
 import { UserServiceInterface } from './user-service.interface.js';
 import UserService from './user.service.js';
-import { AppComponent } from '../../../utils/constants.js';
 import { UserEntity, UserModel } from './user.entity.js';
 import { types } from '@typegoose/typegoose';
+import { AppComponent } from '../../../types/app-component.enum.js';
 
 export function createUserContainer() {
   const container = new Container();
