+++ src/main.ts
@@ -1,11 +1,11 @@
 import { Container } from 'inversify';
 import 'reflect-metadata';
 import Application from './app/application.js';
-import { AppComponent } from './utils/constants.js';
 import { createApplicationContainer } from './app/application.container.js';
 import { createUserContainer } from './core/modules/user/user.container.js';
 import { createLocationContainer } from './core/modules/location/location.container.js';
 import { createOfferContainer } from './core/modules/offer/offer.container.js';
+import { AppComponent } from './types/app-component.enum.js';
 
 async function bootstrap() {
   const container = Container.merge(createApplicationContainer(), createUserContainer(), createLocationContainer(), createOfferContainer());
