+++ src/types/app-component.enum.ts
@@ -1,0 +1,12 @@
+export const AppComponent = {
+  Application: Symbol.for('Application'),
+  LoggerInterface: Symbol.for('LoggerInterface'),
+  ConfigInterface: Symbol.for('ConfigInterface'),
+  DatabaseClientInterface: Symbol.for('DatabaseClientInterface'),
+  UserServiceInterface: Symbol.for('UserServiceInterface'),
+  UserModel: Symbol.for('UserModel'),
+  LocationServiceInterface: Symbol.for('LocationServiceInterface'),
+  LocationModel: Symbol.for('LocationModel'),
+  OfferServiceInterface: Symbol.for('OfferServiceInterface'),
+  OfferModel: Symbol.for('OfferModel'),
+} as const;
