+++ src/types/user.type.ts
@@ -2,5 +2,5 @@ export type User = {
   name: string
   email: string
   avatarUrl: string
-  isPro: boolean
+  userStatus: string
   }
